+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## 6.4.1 (unreleased)
 
 - fix(settings): increase default lifteme of ACCESS and REFRESH token
+- fix(userstories): invalid serializer for userstories if only_ref=true
 
 ## 6.4.0 (2021-09-06)